exit with failure code on prettier error
Zoe should exit with failure code on prettier error.
Not doing so disturb any Git hook library
(like [husky](https://github.com/typicode/husky)), which depend
on the linting to return an error for the hooks to fail or not.

As the way it was Zoe always returned with a sucess code,
precommit were always passing, even when there was errors.
Which is an undesirable effect.